+++ .vscode-test.js
@@ -1,0 +1,4 @@
+// https://code.visualstudio.com/api/working-with-extensions/testing-extension
+const { defineConfig } = require('@vscode/test-cli');
+
+module.exports = defineConfig({ files: 'out/test/**/*.test.js' });